+++ empire/server/plugins/csharpserver/csharpserver.py
@@ -6,17 +6,18 @@ import socket
 import subprocess
 
 import empire.server.common.helpers as helpers
-from empire.server.common.plugins import Plugin
+from empire.server.common.empire import MainMenu
+from empire.server.common.plugins import BasePlugin
+from empire.server.core.db import models
+from empire.server.core.db.base import SessionLocal
+from empire.server.core.db.models import PluginTaskStatus
 from empire.server.core.plugin_service import PluginService
 
 log = logging.getLogger(__name__)
 
 
-class Plugin(Plugin):
+class Plugin(BasePlugin):
     def onLoad(self):
-        self.main_menu = None
-        self.csharpserver_proc = None
-        self.thread = None
         self.info = {
             "Name": "csharpserver",
             "Authors": [
@@ -41,31 +42,42 @@ class Plugin(Plugin):
                 "Strict": True,
             }
         }
+
+        self.csharpserver_proc = None
+        self.thread = None
         self.tcp_ip = "127.0.0.1"
         self.tcp_port = 2012
         self.status = "OFF"
 
     def execute(self, command, **kwargs):
-        try:
-            results = self.do_csharpserver(command)
-            return results
-        except Exception as e:
-            log.error(e)
-            return False, f"[!] {e}"
+        db = kwargs["db"]
+        if command["status"] == "start":
+            input = "Starting Empire C# server..."
+        else:
+            input = "Stopping Empire C# server..."
 
-    def get_commands(self):
-        return self.commands
+        plugin_task = models.PluginTask(
+            plugin_id=self.info["Name"],
+            input=input,
+            input_full=input,
+            user_id=1,
+            status=PluginTaskStatus.completed,
+        )
+        output = self.toggle_csharpserver(command)
+        plugin_task.output = output
+        db.add(plugin_task)
+        db.flush()
 
-    def register(self, mainMenu):
+    def register(self, main_menu: MainMenu):
         """
-        any modifications to the mainMenu go here - e.g.
+        any modifications to the main_menu go here - e.g.
         registering functions to be run by user commands
         """
-        self.installPath = mainMenu.installPath
-        self.main_menu = mainMenu
-        self.plugin_service: PluginService = mainMenu.pluginsv2
+        self.installPath = main_menu.installPath
+        self.main_menu = main_menu
+        self.plugin_service: PluginService = main_menu.pluginsv2
 
-    def do_csharpserver(self, command):
+    def toggle_csharpserver(self, command):
         """
         Check if the Empire C# server is already running.
         """
@@ -101,6 +113,7 @@ class Plugin(Plugin):
                     self.installPath
                     + "/csharp/Covenant/bin/Debug/net6.0/EmpireCompiler.dll",
                 ]
+
                 self.csharpserver_proc = subprocess.Popen(
                     csharp_cmd, stdout=subprocess.PIPE, stderr=subprocess.PIPE
                 )
@@ -118,14 +131,36 @@ class Plugin(Plugin):
                 return "[!] Empire C# server is already started"
 
     def thread_csharp_responses(self):
+        task_input = "Collecting Empire C# server output stream..."
+
         while True:
-            output = self.csharpserver_proc.stdout.readline()
-            if not output:
-                log.warning("csharpserver output stream closed")
-                return
-            output = output.rstrip()
-            if output:
-                log.info(output.decode("UTF-8"))
+            response = self.csharpserver_proc.stdout.readline().rstrip()
+
+            if response:
+                output = response.decode("UTF-8")
+                log.debug(output)
+                status = PluginTaskStatus.completed
+                self.record_task(status, output, task_input)
+
+            elif not response:
+                output = "Empire C# server output stream closed"
+                status = PluginTaskStatus.error
+                log.warning(output)
+                self.record_task(status, output, task_input)
+
+    def record_task(self, status, task_output, task_input):
+        with SessionLocal.begin() as db:
+            plugin_task = models.PluginTask(
+                plugin_id=self.info["Name"],
+                input=task_input,
+                input_full=task_input,
+                user_id=1,
+                status=status,
+            )
+
+            plugin_task.output = task_output
+            db.add(plugin_task)
+            db.flush()
 
     def do_send_message(self, compiler_yaml, task_name, confuse=False):
         bytes_yaml = compiler_yaml.encode("UTF-8")